Add a field to a Vega tooltip-specification

This function is used to add custom fields to a tooltip specification created using vega_tooltip(), according to the options detailed in the Vega tooltip documentation.

add_field(tooltip, field = NULL, title = field, formatType = NULL,
  format = NULL, valueAccessor = NULL, render = NULL, aggregate = NULL,
  ...)

Arguments

tooltip
vega_tooltip object, created using vega_tooltip()

field
character, the unique name of the field. With Vega-Lite, this is the field you provided to each encoding channel.

title
character or JS_EVAL object created using JS(), a custom title for the field, or an accessor function that generates it from the scenegraph datum.

formatType
character, tells what kind of field this is (for formatting the field value in the tooltip) Supported values: "number", "time", and "string".

format
character or JS_EVAL object created using JS()

string: a string specifier for formatting the field value in the tooltip. If formatType is "number", you can provide a number format string-specifier. If formatType is "time", you can provide a time format string-specifier. If formatType is "string", there is no need to provide a format.

function: function that returns a string, in which case formatType is ignored.

valueAccessor
JS_EVAL object created using JS()

render
JS_EVAL object created using JS()

aggregate
character

...
other arguments (not supported)

Value

vega_tooltip object

Details

This function both takes and returns a tooltip specification; it can be used in a "pipeline" that creates a tooltip specification, then adds fields to it.

When making a tooltip specification with custom fields, use the vega_tooltip() function to create the tooltip. This function, when called with its default showAllFields = FALSE, suppresses the default fields, allowing you to propose custom fields.

See also

Examples

  plot_basic <-
    alt$Chart(
      r_to_py(mtcars)
    )$encode(
      x = "mpg:Q",
      y = "hp:Q",
      color = "cyl:N"
    )$mark_point()

  tooltip_custom <-
    vega_tooltip() %>%
    add_field(field = "mpg", title = "MPG") %>%
    add_field(field = "hp", title = "HP")
# NOT RUN {
  vegalite(plot_basic, tooltip = tooltip_custom)
# }