altair: Create and embed Vega-Lite charts using the Altair Python package

The goal of altair is to help you build Vega-Lite visualizations. Using the reticulate package, it provides an interface to the Altair Python package.

Details

In this documentation, the capitalized word Altair shall refer to the Python package; the lower-case word altair shall refer to this R package.

See also