Check the Altair installation

Provides feedback on any differences between your installed version of Altair and the version this package supports.

check_altair(quiet = FALSE)

Arguments

quiet
logical, if TRUE, suppresses message upon successful check

Value

invisible NULL, called for side-effects

Details

If the supported Altair version is different from your installed version, this function will act according to where the difference in the version numbers:

major version leads to an error

minor version leads to a warning

patch version leads to a message

If there is no difference:

quiet = FALSE, success message showing version-numbers

quiet = TRUE, no message

To install the supported version into a Python environment called "r-reticulate", use install_altair().

See also

Examples

# NOT RUN {
  check_altair()
# }