pyr_2010_tit_fpt {altair}	R Documentation

Pyramide des âges des titulaires de la fonction publique territoriale au 31 décembre 2010

Description

Pyramide des âges des titulaires de la fonction publique territoriale au 31 décembre 2010

Usage

pyr_2010_tit_fpt

Format

Une base de données contenant 55 lignes, dont un entête, et 4 variables :

age

âge des agents au 31 décembre 2011

Hommes

nombre d'hommes, en milliers

Femmes

nombre de femmes, en milliers

année.référence

référence de l'année (première ligne)

versant

versant de la fonction publique (FPT ou FPH)

Source

http://www.insee.fr/fr/themes/document.asp?ref_id=ip1442

[Package altair version 0.1 Index]