pyramide_ages {altair}	R Documentation

Pyramide des âges.

Description

Elabore une pyramide des âges verticale avec superposition du début et de la fin de la période sous revue.

Usage

pyramide_ages(Avant, Après = NULL, titre = "",
  date.début = début.période.sous.revue,
  date.fin = fin.période.sous.revue, versant = "",
  couleur_H = "darkslateblue", couleur_F = "firebrick4")

Arguments

Avant
data.table/data.frame décrivant la situation en début de période Cette base doit avoir la forme suivante (bornes d'âges non impératifs):

age 	 Hommes 	 Femmes

15 	 0 	 1

16 	 NA 	 2

17 	 1 	 3

18 	 409 	 52

... 	 ... 	 ...

68 	 2216 	 NA

dans laquelle "age" peut être soit un vecteur de nom de lignes soit une colonne.

Après
data.table/data.frame décrivant la situation en fin de période. Même format que Avant.

titre
Titre du graphique.

date.début
date du début de la période.

date.fin
date de fin de période.

versant
Si non renseigné, sans effet. Si renseigné par "FPT" (resp. "FPH"), le deuxième argument Après ne doit pas être renseigné. Il est automatiquement remplacé par une base de données disponible dans le répertoire data/ du paquet, correspondant à l'année la plus proche du versant de la fonction publique correspondant. La pyramide superposée représente celle qu'aurait l'organisme si la distribution de ses âges était celle du versant mentionné de la fonction publique.

couleur_H
couleur utilisée pour représenter les hommes (partie droite de la pyramide). Par défaut darkslateblue

couleur_F
couleur utilisée pour représenter les femmes (partie gauche de la pyramide). Par défaut firebrick4

Value

Une liste de deux vecteurs numériques représentant chacun des axes (gauche puis droit). Un graphique comprenat une pyramide, une légende et éventuellement un titre.

Examples

pyramide_ages(df1, NULL, "Pyramide des âges", 2008, 2012, versant = "FPT", comparer = TRUE)

[Package altair version 0.1 Index]