Vega datasets

Uses the reticulate framework to access the vega_datasets Python package.

vega_data(...)

Arguments

...
not used

Details

Some data are available locally, others require an internet connection.

See also

Examples

# To list available datasets
vega_data$list_datasets()

#>  [1] "7zip"                            "airports"                       
#>  [3] "anscombe"                        "barley"                         
#>  [5] "birdstrikes"                     "budget"                         
#>  [7] "budgets"                         "burtin"                         
#>  [9] "cars"                            "climate"                        
#> [11] "co2-concentration"               "countries"                      
#> [13] "crimea"                          "disasters"                      
#> [15] "driving"                         "earthquakes"                    
#> [17] "ffox"                            "flare"                          
#> [19] "flare-dependencies"              "flights-10k"                    
#> [21] "flights-200k"                    "flights-20k"                    
#> [23] "flights-2k"                      "flights-3m"                     
#> [25] "flights-5k"                      "flights-airport"                
#> [27] "gapminder"                       "gapminder-health-income"        
#> [29] "gimp"                            "github"                         
#> [31] "graticule"                       "income"                         
#> [33] "iris"                            "jobs"                           
#> [35] "londonBoroughs"                  "londonCentroids"                
#> [37] "londonTubeLines"                 "lookup_groups"                  
#> [39] "lookup_people"                   "miserables"                     
#> [41] "monarchs"                        "movies"                         
#> [43] "normal-2d"                       "obesity"                        
#> [45] "points"                          "population"                     
#> [47] "population_engineers_hurricanes" "seattle-temps"                  
#> [49] "seattle-weather"                 "sf-temps"                       
#> [51] "sp500"                           "stocks"                         
#> [53] "udistrict"                       "unemployment"                   
#> [55] "unemployment-across-industries"  "us-10m"                         
#> [57] "us-state-capitals"               "weather"                        
#> [59] "weball26"                        "wheat"                          
#> [61] "world-110m"                      "zipcodes"                       

# When accessing a dataset, substitute any "-" in the name with a "_"
head(vega_data$sf_temps())

#>   temp                date
#> 1 47.8 2010-01-01 00:00:00
#> 2 47.4 2010-01-01 01:00:00
#> 3 46.9 2010-01-01 02:00:00
#> 4 46.5 2010-01-01 03:00:00
#> 5 46.0 2010-01-01 04:00:00
#> 6 45.8 2010-01-01 05:00:00

# Metadata are available for each dataset:
vega_data$sf_temps$description

#> [1] "This dataset contains hourly temperatures in San Francisco during the full year of 2010. The dataset is drawn from public-domain `NOAA data <https://www.weather.gov/disclaimer>`_, and transformed using scripts available at http://github.com/vega/vega_datasets/."

vega_data$sf_temps$url

#> [1] "https://vega.github.io/vega-datasets/data/sf-temps.csv"

# For local datasets, local path is available
vega_data$sf_temps$filepath

#> [1] "/anaconda3/envs/r-reticulate/lib/python3.6/site-packages/vega_datasets/data/sf-temps.csv"