Vega embed options

Use this function to specify embedding-options to the vegalite renderer. The most common options to set are renderer, to specify "canvas" or "svg", and actions, to specify the inclusion of export, source, and editor links. This documentation is adapted from the vega-embed documentation.

vega_embed(renderer = c("canvas", "svg"), actions = TRUE, mode = NULL,
  logLevel = NULL, loader = NULL, onBeforeParse = NULL, width = NULL,
  height = NULL, padding = NULL, config = NULL, editorUrl = NULL,
  sourceHeader = NULL, sourceFooter = NULL, runAsync = NULL)

Arguments

renderer
character the renderer to use for the view. One of "canvas" (default) or "svg". See Vega docs for details.

actions
logical or named vector of logicals, determines if action links ("Export as PNG/SVG", "View Source", "Open in Vega Editor") are included with the embedded view. If the value is TRUE (default), all action links will be shown and none if the value is FALSE. This property can be a named vector of logicals that maps keys (export, source, editor) to logical values for determining if each action link should be shown. Unspecified keys will be true by default. For example, if actions is list(export = FALSE, source = TRUE), the embedded visualization will have two links – "View Source" and "Open in Vega Editor".

mode
character if specified, tells Vega-Embed to parse the spec as vega or vega-lite. Vega-Embed will parse the $schema url if the mode is not specified. Vega-Embed will default to vega if neither mode, nor $schema are specified.

logLevel
sets the current log level. See Vega docs for details.

loader
sets a custom Vega loader. See Vega docs for details.

onBeforeParse
JS function, modifies the spec before being parsed.

width
integer sets the view width in pixels. See Vega docs for details. Note that Vega-Lite overrides this option.

height
integer, sets the view height in pixels. See Vega docs for details. Note that Vega-Lite overrides this option.

padding
list, sets the view padding in pixels. See Vega docs for details.

config
character or list a URL string** from which to load a Vega/Vega-Lite or Vega-Lite configuration file, or a list of Vega/Vega-Lite configurations to override the default configuration options. If config is a URL, it will be subject to standard browser security restrictions. Typically this URL will point to a file on the same host and port number as the web page itself.

editorUrl
character, URL at which to open embedded Vega specs in a Vega editor. Defaults to "http://vega.github.io/editor/". Internally, Vega-Embed uses HTML5 postMessage to pass the specification information to the editor.

sourceHeader
character, HTML to inject into the `` tag of the page generated by the "View Source" action link. For example, this can be used to add code for syntax highlighting.

sourceFooter
character, HTML to inject into the end of the page generated by the "View Source" action link. The text will be added immediately before the closing `` tag.

runAsync
logical, indicating to use runAsync instead of run.

Value

S3 object with class vega_embed

Details

The default renderer is "canvas".

The default for actions is TRUE, which means that all three action links will be included.

To suppress all action links, call with actions = FALSE.

To suppress a given action link, call with a list: actions = list(editor = FALSE).

To specify which link(s) to include, call using the only_actions() function: actions = only_actions(export = TRUE).

It is ineffective to set the width and height parameters here, as they will be overridden by the values in the chart specification.

See also

Examples

# Set the renderer
embed_options <- vega_embed(renderer = "svg")

# Specify the action links
embed_options <- vega_embed(actions = FALSE)
embed_options <- vega_embed(actions = list(editor = FALSE))
embed_options <- vega_embed(actions = only_actions(export = TRUE))