Create a Vega tooltip-specification

This function creates a Vega tooltip-specification according to the options detailed in the Vega tooltip documentation.

vega_tooltip(showAllFields = FALSE, fields = NULL, delay = 100,
  onAppear = NULL, onMove = NULL, onDisappear = NULL,
  colorTheme = c("light", "dark"), sort = c("title", "value"))

vega_tooltip_encoding(showAllFields = NULL, ...)

vega_tooltip_all(showAllFields = TRUE, ...)

Arguments

showAllFields
logical, indicates which fields to include in tooltip.

TRUE: include all fields in the data.

FALSE: include only those fields specified in fields. Use this with empty fields to suppress tooltip.

NULL: include only those fields encoded in the chart.

fields
list, tooltip specification to use if showAllFields == FALSE. Use add_field() to specify.

delay
numeric number of milliseconds tooltip display should be delayed.

onAppear
JS_EVAL object created using JS(), a function with arguments: event, item. Callback when tooltip first appears (when mouse moves over a new item in visualization).

onMove
JS_EVAL object created using JS(), a function with arguments: event, item. Callback when tooltip moves (e.g., when mouse moves within a bar).

onDisappear
JS_EVAL object created using JS(), a function with arguments: event, item. Callback when tooltip disappears (when mouse moves out an item).

colorTheme
character color theme picker

sort
character or JS_EVAL object created using htmlwidgets::JS(), specify sorting fields on either "title", or "value", using sorting orders:

dates and strings are sorted ascending, numbers descending

dates appear first, then numbers, then strings

...
other arguments passed to vega_tooltip()

Value

S3 object with class vega_tooltip

Details

This function is called vega_tooltip() because the tooltip specification made for the Vega JavaScript library; it is independent from the Altair Python package.

A tooltip specification is separate from a chart specification; they are combined when they are both rendered into an htmlwidget, using vegalite(). One conseqence of this is that you can have exactly one tooltip specification per chart, even a compound chart.

There are three versions of this function, each will result in a different default behavior of the tooltip:

vega_tooltip()

no tooltip is displayed, absent the use of add_field()

vega_tooltip_encoding()

tooltip displays all the variables in the chart's encoding

vega_tooltip_all()

tooltip displays all the variables in the chart's data

To customize the fields in a tooltip specification, begin with with a specification with no fields, using vega_tooltip(). To add fields (variables) to the specification, pipe this to the add_field() function for as many fields as you wish to add. See the included example that composes tooltip_custom.

You can use the examine() function to interactively examine a tooltip specification.

See also

Examples

  plot_basic <-
    alt$Chart(
      r_to_py(mtcars)
    )$encode(
      x = "mpg:Q",
      y = "hp:Q",
      color = "cyl:N"
    )$mark_point()
# NOT RUN {
  vegalite(plot_basic, tooltip = vega_tooltip_encoding())
# }

tooltip_custom <-
  vega_tooltip() %>%
  add_field(field = "mpg", title = "MPG") %>%
  add_field(field = "hp", title = "HP")
# NOT RUN {
  vegalite(plot_basic, tooltip = tooltip_custom)
# }