Create a Vega/Vega-Lite htmlwidget

Renders a chart as an htmlwidget.

vegawidget(chart, embed = NULL, width = NULL, height = NULL, ...)

Arguments

chart
an Altair chart object

embed
vega_embed object to specify embedding options - the default is an empty call to vega_embed(), which will result in a canvas-rendering and action-links included for exporting, viewing the Vega-Lite source, and opening the Vega editor.

width
integer, if specified, the total rendered width (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the width in the chart specification

height
integer, if specified, the total rendered height (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the height in the chart specification

...
other arguments

Details

This function is called vegawidget() is because it returns an htmlwidget that uses Vega-Lite and Vega JavaScript libraries, rather than the Altair Python package.

To specify embedding-options, use the vega_embed() function with the embed argument. Its most-important options are:

renderer, to specify "canvas" (default) or "svg"

actions, to specify action-links for export, source, compiled, and editor

If actions is TRUE (or FALSE), all the links are shown (or not). Use a named list to be more specific, see vega_embed() and only_actions().

If embed is NULL, vegawidget() sets embed to the value of getOption("altair.embed_options"). Then, if this option is NULL, the vega-embed defaults are used.

The arguments width and height are used to override the width and height determined using the chart specification. However, there are some important provisions:

Specifying width and height in vegawidget() is effective only for single-view charts and layered charts. It will not work for contatenated, faceted, or repeated charts.

In the chart specification, the default interpretation of width and height is to describe the dimensions of the plotting rectangle, not including the space used by the axes, labels, etc. When width and height are specified using vegawidget(), the meanings change to describe the dimensions of the entire rendered chart, including axes, labels, etc.

Keep in mind that the action-links are not a part of the rendered chart, so you may have to account for them yourself. You might expect the height of the action-links to be 15-20 pixels.

See also

Examples

vega_data <- import_vega_data()

chart <-
  alt$Chart(r_to_py(vega_data$cars()))$
  mark_point()$
  encode(
    x = "Horsepower:Q",
    y = "Miles_per_Gallon:Q",
    color = "Origin:N"
  )
# NOT RUN {
  # default: rendered using canvas,
  #  default action-links, chart-specification determines size
  vegawidget(chart)

  # render using SVG
  vegawidget(chart, embed = vega_embed(renderer = "svg"))

  # do not include action-links
  vegawidget(chart, embed = vega_embed(actions = FALSE))

  # specify dimensions of rendered-chart
  vegawidget(chart, width = 300, height = 200)
# }