Shiny output for Vega-Lite

Shiny output for Vega-Lite

vegawidgetOutput(outputId, width = "100%", height = "400px")

Arguments

outputId
output variable to read from

width
Must be a valid CSS unit (like "100%", "400px", "auto") or a number, which will be coerced to a string and have "px" appended.

height
Must be a valid CSS unit (like "100%", "400px", "auto") or a number, which will be coerced to a string and have "px" appended.