Install Altair Python package

This function wraps installation functions from reticulate to install the Python packages altair and vega_datasets.

install_altair(
  method = c("conda", "virtualenv"),
  envname = "r-reticulate",
  version = getOption("altair.python.version"),
  ...
)

Arguments

method

character, indicates to use "conda" or "virtualenv"

envname

character, name of environment into which to install

version

character, version of Altair to install. For general use of this package, this is set automatically, so you should not need to specify this.

...

other arguments sent to reticulate::py_install()

Value

invisible NULL, called for side-effects

Details

This package uses the reticulate package to make an interface with the Altair Python package. To promote consistency in usage of reticulate among different R packages, it is recommended to use a common Python environment, called "r-reticulate".

Depending on your setup, you can create this environment using reticulate::conda_create() or reticulate::virtualenv_create(), as described in this reticulate article, or in this package's Installation article.

See also

Examples

if (FALSE) {
  # not run because it requires Python
  install_altair()
}