Create a Vega/Vega-Lite htmlwidget

Use this function to render a chart as an htmlwidget.

vegawidget(chart, embed = NULL, width = NULL, height = NULL, ...)

Arguments

chart
an Altair plot object

embed
vega_embed object to specify embedding options - the default is an empty call to vega_embed(), which will result in a canvas-rendering and all actions-links being included

width
integer, if specified, the total rendered width (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the chart specification

height
integer, if specified, the total rendered height (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the chart specification

...
other arguments

Details

This function is called vegawidget() is because it returns an htmlwidget that uses the Vega-Lite JavaScript library, rather than the Altair Python package.

To specify embedding-options, use the vega_embed() function with the embed argument. Its most-important options are:

renderer, to specify "canvas" (default) or "svg"

actions, to specify action-links for export, source, and editor

If actions is TRUE (default) or FALSE, all the links are shown (or not). Use a named list to be more specific, see vega_embed() and only_actions().

If embed is NULL, vegawidget() sets embed to the value of getOption("altair.embed_options"). If that is null, the vega-embed defaults are used.

The arguments width and height are used to override the width and height determined using the chart specification. However, there are some important provisions:

Specifying width and height in vegawidget() is effective only for single-view charts and layered charts. It will not work for contatenated, faceted, or repeated charts.

In the chart specification, the default interpretation of width and height is to describe the dimensions of the plotting rectangle, not including the space used by the axes, labels, etc. When width and height are specified using vegawidget(), the meanings change to describe the dimensions of the entire rendered chart, including axes, labels, etc.

Keep in mind that the action-links are not a part of the rendered chart, so you may have to account for them yourself. You might expect the height of the action-links to be 25-30 pixels.

See also

Examples

  plot_basic <-
    alt$Chart(
      r_to_py(mtcars)
    )$encode(
      x = "mpg:Q",
      y = "hp:Q",
      color = "cyl:N"
    )$mark_point()
# NOT RUN {
  # default: rendered using canvas,
  #   all action-links, chart-specification determines size
  vegawidget(plot_basic)

  # render using SVG
  vegawidget(plot_basic, embed = vega_embed(renderer = "svg"))

  # do not include action-links
  vegawidget(plot_basic, embed = vega_embed(actions = FALSE))

  # specify dimensions of rendered-chart
  vegawidget(plot_basic, width = 300, height = 200)
# }